requirements.txt
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
crispy-forms==2.4
crypto==1.4.1
dj-rest-auth==7.0.1
Django==5.1.4
django-crispy-forms==2.4
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
django-rest-auth==7.0.1
oauthlib==4.2.1
Pillow==11.1.0
python-dotenv==1.0.0